requirements.txt
asgiref==3.8.1
azure-core==1.32.0
azure-storage-blob==12.24.1
beautifulsoup4==4.12.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary
cryptography==44.0.1
Django
django-cloudinary-storage
django-livereload==1.7
django-livereload-server==0.5.1
django-livesync==0.5
django-storages==1.14.5
django-user-agents==0.4.0
idna==3.10
isodate==0.7.2
livereload==2.7.1
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
qrcode==8.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
watchdog==6.0.0
websocket-client==1.8.0
whitenoise==6.9.0
gunicorn
